Collect sample_book and sample_paper into separate module.
